requirements.txt
streamlit
rapidocr
Pillow
numpy